ollama_service/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic>=2.5.3,<3.0.0
ollama-python==0.1.2
